Rearranged the files, separating more code
