+++ trombi/static/js/lang_calendar/fr.js
